pyproject.toml
[project]
name = "langchain-hw"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.31",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
]
